Dependecy Fix
- Cyclic import fixed
- Library utilities and model import
- view.py removed
