+++ lign/utils/io.py
@@ -1,8 +1,6 @@
 import json as jn
 import pickle as pk
 import os
-from torchvision import datasets
-import torch
 from collections.abc import Iterable
 
 def unpickle(fl):
@@ -36,39 +34,3 @@ def to_iter(data):
 
 def is_primitve(data):
     return type(data) in (int, str, bool, float)
-
-def cal_to_lign(path, transforms = None, train = True):
-    pass
-
-def cifar_to_lign(path, transforms = None, train = True):
-    from ..graph import GraphDataset
-
-    dataset =  datasets.CIFAR100(path, train=train)
-    graph = GraphDataset()
-
-    imgs = []
-    labels = []
-    for img, lab in dataset:
-        out = {
-                "data": {},
-                "edges": set()
-            }
-
-        if transforms:
-            img = transforms(img)
-
-        imgs.append(img)
-        labels.append(lab)
-        graph.add(out)
-    
-    if(torch.is_tensor(imgs[0])):
-        imgs = torch.stack(imgs)
-        labels = torch.LongTensor(labels)
-
-    graph.set_data('x', imgs)
-    graph.set_data('labels', labels)
-
-    return graph
-
-def core_to_lign(path, transforms = None, train = True):
-    pass
