+++ server-training/cifar.py
@@ -13,8 +13,8 @@ import numpy as np
 import datetime
 tm_now = datetime.datetime.now
 
-dataset = lg.graph.GraphDataset("data/datasets/cifar100_train.lign")
-validate = lg.graph.GraphDataset("data/datasets/cifar100_test.lign")
+dataset = lg.graph.GraphDataset("../data/datasets/cifar100_train.lign")
+validate = lg.graph.GraphDataset("../data/datasets/cifar100_test.lign")
 
 if th.cuda.is_available():
     device = th.device("cuda")
@@ -112,7 +112,7 @@ metrics = {
     "accuracy": accuracy,
     "log": log
 }
-utl.io.json(metrics, "data/metrics/"+filename+".json")
+utl.io.json(metrics, "../data/metrics/"+filename+".json")
 
 ## Save hyperparameters
 para = {
